chore: move mempool sync state machine into its own file
